examples/requirements.txt
# Python dependencies for Pamir AI E-Ink examples
# Install with: pip install -r requirements.txt

# Image processing library for creating graphics
Pillow>=9.0.0

# HTTP requests for weather dashboard
requests>=2.28.0

# Optional: For better font support
# freetype-py>=2.3.0

# For GIF animation support
imageio[ffmpeg]>=2.31.0

# Optional: For advanced graphics
# numpy>=1.21.0
# matplotlib>=3.5.0